fix: change to named imports

+++ packages/foundry/contracts/UniswapV2.sol
@@ -10,10 +10,7 @@ contract UniswapV2PriceChecker {
     address constant WETH = 0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2;
     address constant USDC = 0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48;
 
-    function getAmountsOut(
-        uint amountIn,
-        address[] memory path
-    ) public view returns (uint[] memory) {
+    function getAmountsOut(uint256 amountIn, address[] memory path) public view returns (uint256[] memory) {
         return IUniswapV2Router(ROUTER).getAmountsOut(amountIn, path);
     }
 }
